feat: normalize on minimum field value
The heuristic is admissible so long as it is a consistent
underestimate of the full path length. If the distance measure
is too small ( h(x) << g(x) ), then it may have little influence
over how path selection is prioritized ( h(x) = 0 => ordinary dijkstra ).

We can accomodate more diverse data by normalizing based on the smallest
value present in the data (which are understood to be all positive values).