feat: add A* search via "compass" parameter to djikstra (#6)
* feat: add A* search via "line" heuristic parameter to djikstra

A* ("best first") search can use certain admissible heuristics
to improve search speed. We can use the straight line distance
to target to improve search efficiency without sacrificing memory
usage.

This update adds:

   dijkstra.dijkstra(..., heuristic='line') # A* search
   dijkstra.dijkstra(..., heuristic=None) # ordinary dijkstra

This improves search speed without sacrificing memory. Despite requiring
more FLOPS, most of the latency in the algorithm is due to memory latency
when fetching across Z.

In practice, this update looks like it improves soma processing by about
2x and has negligable impact on ordinary neurite tracing.

* fix+perf: fixing implementation of A*

* fix: working on making sure this is an admissible heuristic

The squared distance rapidly increases and probably violates the
admissibility constraint. Trying to determine if the straight line
distance is admissible for the PDRF.

* perf: use Chebychev (chessboard) distance heuristic

* test: A* should find an equal length path but maybe not exactly the same

* feat: normalize on minimum field value

The heuristic is admissible so long as it is a consistent
underestimate of the full path length. If the distance measure
is too small ( h(x) << g(x) ), then it may have little influence
over how path selection is prioritized ( h(x) = 0 => ordinary dijkstra ).

We can accomodate more diverse data by normalizing based on the smallest
value present in the data (which are understood to be all positive values).

* test: allow all viable solutions for one of the tests

astar was picking one of the other solutions

* feat: allow custom normalizer for a*

* feat: python interface for selecting custom normalizer

* feat: respect zero weight paths

* fix: heuristic was being computed incorrectly

Accidently computing the heuristic for the same location
all the time instead of for the neighbor we should have been
considering.

* refactor: make the xyz computation less repetitive